+++ CHANGELOG.rst
@@ -1,8 +1,8 @@
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-Changelog for package bcr_bot
+Changelog for package amr_mtt
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 1.0.2 (2024-04-02)
 ------------------
 * Setting Gazebo resource paths through launch files
-* Modified launch files to dynamically spawn bcr_bot
+* Modified launch files to dynamically spawn amr_mtt